requirements.txt
web3
flask==2.0
python-dotenv